move BranchPicker to contrib/chat/browser for cross-contrib reuse
BranchPicker needs to be importable by non-provider contribs (e.g.
automations). Moving it from contrib/providers/copilotChatSessions
to contrib/chat/browser satisfies the sessions layer rules.

Copilot-Session: 0d30007f-3037-4015-a850-60e8d3852d66
