+++ src/vs/sessions/contrib/chat/browser/branchPicker.ts
@@ -3,25 +3,24 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as dom from '../../../../../base/browser/dom.js';
-import { Gesture, EventType as TouchEventType } from '../../../../../base/browser/touch.js';
-import { Codicon } from '../../../../../base/common/codicons.js';
-import { Disposable, DisposableStore } from '../../../../../base/common/lifecycle.js';
-import { autorun, IObservable } from '../../../../../base/common/observable.js';
-import { localize } from '../../../../../nls.js';
-import { IActionWidgetService } from '../../../../../platform/actionWidget/browser/actionWidget.js';
-import { ActionListItemKind, IActionListDelegate, IActionListItem } from '../../../../../platform/actionWidget/browser/actionList.js';
-import { ITelemetryService } from '../../../../../platform/telemetry/common/telemetry.js';
-import { renderIcon } from '../../../../../base/browser/ui/iconLabel/iconLabels.js';
-import { reportNewChatPickerClosed } from '../../../chat/browser/newChatPickerTelemetry.js';
+import * as dom from '../../../../base/browser/dom.js';
+import { Gesture, EventType as TouchEventType } from '../../../../base/browser/touch.js';
+import { Codicon } from '../../../../base/common/codicons.js';
+import { Disposable, DisposableStore } from '../../../../base/common/lifecycle.js';
+import { autorun, IObservable } from '../../../../base/common/observable.js';
+import { localize } from '../../../../nls.js';
+import { IActionWidgetService } from '../../../../platform/actionWidget/browser/actionWidget.js';
+import { ActionListItemKind, IActionListDelegate, IActionListItem } from '../../../../platform/actionWidget/browser/actionList.js';
+import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry.js';
+import { renderIcon } from '../../../../base/browser/ui/iconLabel/iconLabels.js';
+import { reportNewChatPickerClosed } from './newChatPickerTelemetry.js';
 
 const FILTER_THRESHOLD = 10;
 
 /** Minimal contract the BranchPicker needs to render and interact. */
 export interface IBranchPickerModel {
 	readonly branches: IObservable<readonly string[]>;
-	readonly branch: IObservable<string | undefined>;
-	readonly loading: IObservable<boolean>;
+	readonly selectedBranch: IObservable<string | undefined>;
 	readonly disabled: IObservable<boolean>;
 	setBranch(name: string): void;
 }
@@ -50,9 +49,8 @@ export class BranchPicker extends Disposable {
 		this._register(autorun(reader => {
 			const model = this._model.read(reader);
 			if (model) {
-				model.loading.read(reader);
 				model.branches.read(reader);
-				model.branch.read(reader);
+				model.selectedBranch.read(reader);
 				model.disabled.read(reader);
 			}
 			this._updateTriggerLabel();
@@ -95,7 +93,7 @@ export class BranchPicker extends Disposable {
 			return;
 		}
 
-		const selectedBranch = model?.branch.get();
+		const selectedBranch = model?.selectedBranch.get();
 		const items: IActionListItem<IBranchItem>[] = branches.map(branch => ({
 			kind: ActionListItemKind.Action,
 			label: branch,
@@ -146,9 +144,8 @@ export class BranchPicker extends Disposable {
 		dom.clearNode(this._triggerElement);
 
 		const model = this._model.get();
-		const isLoading = model?.loading.get() ?? false;
-		const isDisabled = model?.disabled.get() ?? false;
-		const label = model?.branch.get() ?? localize('branchPicker.select', "Branch");
+		const isDisabled = model?.disabled.get() ?? true;
+		const label = model?.selectedBranch.get() ?? localize('branchPicker.select', "Branch");
 
 		dom.append(this._triggerElement, renderIcon(Codicon.gitBranch));
 		const labelSpan = dom.append(this._triggerElement, dom.$('span.sessions-chat-dropdown-label'));
@@ -157,8 +154,8 @@ export class BranchPicker extends Disposable {
 
 		this._triggerElement.ariaLabel = localize('branchPicker.triggerAriaLabel', "Pick Branch, {0}", label);
 
-		this._slotElement?.classList.toggle('disabled', isLoading || isDisabled);
-		this._triggerElement.setAttribute('aria-disabled', String(isLoading || isDisabled));
-		this._triggerElement.tabIndex = (isLoading || isDisabled) ? -1 : 0;
+		this._slotElement?.classList.toggle('disabled', isDisabled);
+		this._triggerElement.setAttribute('aria-disabled', String(isDisabled));
+		this._triggerElement.tabIndex = isDisabled ? -1 : 0;
 	}
 }
